# [counter] Use an atomic swap when reporting a counter