@prashantv 2016-10-23
While the current interface only supports reporting, I think we should also have a `snapshot` method that lets you access the current value of the counter. This would be useful for having a debug endpoint that exposes the stats from this specific instance.

I don't think having `report` be goroutine safe needs to be a goal. Report is intended to be called periodically, so there should only really be one goroutine that does this.