@jeromefroe 2016-10-23
While I agree that the intended use of `report` doesn't require it be goroutine safe, I think it would be preferable if we could make it goroutine safe and support a `snapshot` function. To that end, I think we could do the following:

```
type counter struct {
	total int64
	curr  int64
}

func (c *counter) Inc(v int64) {
	atomic.AddInt64(&c.curr, v)
}

func (c *counter) report(name string, tags map[string]string, r StatsReporter) {
	curr := atomic.SwapInt64(&c.curr, 0)

	if curr == 0 {
		return
	}
	atomic.AddInt64(&c.total, curr)
	r.ReportCounter(name, tags, curr)
}
```

With this implementation the size of a `counter` stays the same size as it is now, and `report` should  also be just as fast since it is using only two atomic instructions as well. We could then implement `snapshot` as follows:

```
func (c *counter) snapshot() int64 {
        total := atomic.LoadInt64(&c.total)
        curr := c.curr
        return total + curr
}
```

This implementation of `snapshot` may lag behind the "true" count when `report` sleeps after calling `curr := atomic.SwapInt64(&c.curr, 0)` but before `atomic.AddInt64(&c.total, curr)` and `snapshot` is called in between. However, I think this is ok if `snapshot` is used for debugging purposes since it would only occur very rarely.