@jeromefroe 2016-11-22
After @robskillington's changes to `gauge` this pull request no longer makes sense so I'm going to close it.